Sync code with existing Toast bar
