+++ packages/ui/src/App.tsx
@@ -3,7 +3,7 @@ import { AccountContextProvider } from './contexts/AccountsContext'
 import { ApiContextProvider } from './contexts/ApiContext'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { MultiProxyContextProvider } from './contexts/MultiProxyContext'
-import { ToastContextProvider } from './contexts/ToastContext'
+import ToastContextProvider from './contexts/ToastContext'
 import { theme } from './theme'
 import { ThemeProvider as MuiThemeProvider } from '@mui/material/styles'
 import { AccountNamesContextProvider } from './contexts/AccountNamesContext'
