+++ packages/ui/src/components/Snackbar.tsx
@@ -1,13 +1,13 @@
 import { Stack, Snackbar as MuiSnackbar, styled } from "@mui/material";
-import { useSnackStack } from "./SnackStackProvider";
-import ToastBar from "../ToastBar";
+import { useToasts } from "../contexts/ToastContext";
+import ToastBar from "./ToastBar";
 
 interface Props {
   className?: string
 }
 
 const Snackbar: React.FC  = ({ className }: Props) => {
-  const { toastsPack } = useSnackStack();
+  const { toastsPack } = useToasts();
   const firstToast = toastsPack[0];
 
   return (
@@ -15,7 +15,6 @@ const Snackbar: React.FC  = ({ className }: Props) => {
       className={className}
       open={!!firstToast}
       autoHideDuration={12000}
-      transitionDuration={0} 
       anchorOrigin={{
         vertical: firstToast?.position?.vertical || "bottom",
         horizontal: firstToast?.position?.horizontal || "left"
